Guard added
You can now use Guard for rapid development.
While the integration test via rake starts his own ArangoDB instance,
guard will not, because we want to keep it running and only execute
the integration tests, which kills the overhead of restarting the
instance all the time.
The tests run really, really fast.

In order to achieve that, I outsourced the ArangoDB starter into
a separate file, which is used by the rake task.